+++ src/pcm_conv.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 #include "pcm.h"
 
-size_t pcm_float_to_fixed(audio_format_t af, uint8_t* dst, float* src, size_t frames);
-size_t pcm_fixed_to_float(audio_format_t af, float* dst, uint8_t* src, size_t frames);
+uint32_t pcm_float_to_fixed(audio_format_t af, uint8_t* dst, float* src, uint32_t frames);
+uint32_t pcm_fixed_to_float(audio_format_t af, float* dst, uint8_t* src, uint32_t frames);
 
 #endif
